Added a method to allow the update of both the current and the total for the pice chart so it's display remains overall accurate.
- (void)updateChartByCurrent:(NSNumber *)current byTotal:(NSNumber *)total;

We have a scenario where our app is pulling updates to both the "checked in" (our current) and "registered and ready to check in" (our total). Since both values can be updated, the chart would not be correct if only the option for the total to be updated was not available.